CORE-2570 Migrate ro realm-kotlin

+++ infra/realm/src/main/java/com/simprints/infra/realm/RealmWrapperImpl.kt
@@ -4,49 +4,56 @@ import android.content.ComponentName
 import android.content.Context
 import android.content.Intent
 import android.os.Build
-import com.simprints.infra.logging.LoggingConstants.CrashReportTag.*
-import com.simprints.infra.logging.Simber
 import com.simprints.infra.authstore.AuthStore
+import com.simprints.infra.logging.LoggingConstants.CrashReportTag.DB_CORRUPTION
+import com.simprints.infra.logging.LoggingConstants.CrashReportTag.REALM_DB
+import com.simprints.infra.logging.Simber
 import com.simprints.infra.realm.config.RealmConfig
 import com.simprints.infra.realm.exceptions.RealmUninitialisedException
 import com.simprints.infra.security.SecurityManager
 import com.simprints.infra.security.keyprovider.LocalDbKey
 import dagger.hilt.android.qualifiers.ApplicationContext
-import io.realm.Realm
-import io.realm.RealmConfiguration
-import io.realm.exceptions.RealmFileException
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
+import io.realm.kotlin.MutableRealm
+import io.realm.kotlin.Realm
+import io.realm.kotlin.RealmConfiguration
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
 class RealmWrapperImpl @Inject constructor(
     @ApplicationContext private val appContext: Context,
+    private val configFactory: RealmConfig,
     private val securityManager: SecurityManager,
-    private val authStore: AuthStore
+    private val authStore: AuthStore,
 ) : RealmWrapper {
 
     private lateinit var config: RealmConfiguration
 
     private fun initRealm() {
         if (!this::config.isInitialized) {
-            Realm.init(appContext)
             config = createAndSaveRealmConfig()
         }
     }
 
-    /**
-     * Use realm instance in from IO threads
-     * throws RealmUninitialisedException
-     */
-    override suspend fun <R> useRealmInstance(block: (Realm) -> R): R =
-        withContext(Dispatchers.IO) {
-            initRealm()
-            Simber.tag(REALM_DB.name).d("[RealmWrapperImpl] getting new realm instance")
+    private fun getRealm(): Realm {
+        initRealm()
+        Simber.tag(REALM_DB.name).d("[RealmWrapperImpl] getting new realm instance")
+        return try {
             try {
-                Realm.getInstance(config).use(block)
-            } catch (ex: RealmFileException) {
+                Realm.open(config)
+            } catch (ex: IllegalStateException) {
+                // On the schema update realm it is not being closed correctly and there
+                // is no legitimate way to forcefully restart the realm instance,
+                // so we need to catch this exception and try opening realm again.
+                // If the exception repeats, it should be propagated up the call stack.
+                if (!isFileCorruptException(ex)) {
+                    Realm.open(config)
+                } else {
+                    throw ex
+                }
+            }
+        } catch (ex: Exception) {
+            if (isFileCorruptException(ex)) {
                 //DB corruption detected; either DB file or key is corrupt
                 //1. Delete DB file in order to create a new one at next init
                 Realm.deleteRealm(config)
@@ -59,13 +66,38 @@ class RealmWrapperImpl @Inject constructor(
                 //5. Delete "last sync" info and start new sync
                 resetDownSyncState()
                 //6. Retry operation with new file and key
-                Realm.getInstance(config).use(block)
+                Realm.open(config)
+            } else {
+                throw ex
             }
         }
+    }
+
+    private fun isFileCorruptException(ex: Exception) = ex is IllegalStateException
+        && ex.message?.contains("RLM_ERR_INVALID_DATABASE") == true
+
+    /**
+     * Executes provided block ensuring a valid Realm instance is used and closed.
+     */
+    override suspend fun <R> readRealm(block: (Realm) -> R): R {
+        val realm = getRealm()
+        val result = block(realm)
+        realm.close()
+        return result
+    }
+
+    /**
+     * Executes provided block in a transaction ensuring a valid Realm instance is used and closed.
+     */
+    override suspend fun <R> writeRealm(block: (MutableRealm) -> R) {
+        val realm = getRealm()
+        realm.write(block)
+        realm.close()
+    }
 
     private fun createAndSaveRealmConfig(): RealmConfiguration {
         val localDbKey = getLocalDbKey()
-        return RealmConfig.get(localDbKey.projectId, localDbKey.value, localDbKey.projectId)
+        return configFactory.get(localDbKey.projectId, localDbKey.value)
     }
 
     private fun getLocalDbKey(): LocalDbKey =
